+++ src/core/index.ts
@@ -41,3 +41,4 @@ export {
 
 export * from './profile'
 export * from './learning'
+export * from './priceIntelligence'